project_rag/requirements.txt
openai
python-dotenv
sentence-transformers
torch